:children_crossing: lock-screen: check inactivity after locked
Right now a user can lock the screen either by inactivity (after 600 seconds) or by forcing it using a hot corner or by calling the script with --force. The problem is that if the user comes back and moves the mouse, the screen turns on but never turns off again due to inactivity (in case the user goes away again without unlocking).
This adds a new timer of 30s to xidlehook, which will run lock-screen
--recheck-inactivity. Essentially, it will check if i3lock is still
running and if that's the case it will turn off the screen. This runs
once every 30 seconds of inactivity.